requirements.txt
pandas
streamlit
numpy
mysql-connector-python
streamlit-aggrid